Move OpenSuse to DistributionInstaller
